// src/playfab_sdk/playfab.types.ts
export interface CharacterInventory {
    /** The id of this character. */
    CharacterId?: string;
    /** The inventory of this character. */
    Inventory?: ItemInstance[];

}
// ...
export interface CharacterResult {
    /** The id for this character on this player. */
    CharacterId?: string;
    /** The name of this character. */
    CharacterName?: string;
    /** The type-string that was given to this character on creation. */
    CharacterType?: string;

}
// ...
export interface EntityKey {
    /** Unique ID of the entity. */
    Id: string;
    /** Entity type. See https://docs.microsoft.com/gaming/playfab/features/data/entities/available-built-in-entity-types */
    Type?: string;

}

export interface EntityTokenResponse {
    /** The entity id and type. */
    Entity?: EntityKey;
    /** The token used to set X-EntityToken for all entity based API calls. */
    EntityToken?: string;
    /** The time the token will expire, if it is an expiring token, in UTC. */
    TokenExpiration?: string;

}
// ...
export interface GetPlayerCombinedInfoResult extends PlayFabModule.IPlayFabResultCommon  {
    /** Results for requested info. */
    InfoResultPayload?: GetPlayerCombinedInfoResultPayload;
    /** Unique PlayFab assigned ID of the user on whom the operation will be performed. */
    PlayFabId?: string;

}

export interface GetPlayerCombinedInfoResultPayload {
    /** Account information for the user. This is always retrieved. */
    AccountInfo?: UserAccountInfo;
    /** Inventories for each character for the user. */
    CharacterInventories?: CharacterInventory[];
    /** List of characters for the user. */
    CharacterList?: CharacterResult[];
    /**
     * The profile of the players. This profile is not guaranteed to be up-to-date. For a new player, this profile will not
     * exist.
     */
    PlayerProfile?: PlayerProfileModel;
    /** List of statistics for this player. */
    PlayerStatistics?: StatisticValue[];
    /** Title data for this title. */
    TitleData?: { [key: string]: string | null };
    /** User specific custom data. */
    UserData?: { [key: string]: UserDataRecord };
    /** The version of the UserData that was returned. */
    UserDataVersion: number;
    /** Array of inventory items in the user's current inventory. */
    UserInventory?: ItemInstance[];
    /** User specific read-only data. */
    UserReadOnlyData?: { [key: string]: UserDataRecord };
    /** The version of the Read-Only UserData that was returned. */
    UserReadOnlyDataVersion: number;
    /** Dictionary of virtual currency balance(s) belonging to the user. */
    UserVirtualCurrency?: { [key: string]: number };
    /** Dictionary of remaining times and timestamps for virtual currencies. */
    UserVirtualCurrencyRechargeTimes?: { [key: string]: VirtualCurrencyRechargeTime };

}
// ...
export interface ItemInstance {
    /** Game specific comment associated with this instance when it was added to the user inventory. */
    Annotation?: string;
    /** Array of unique items that were awarded when this catalog item was purchased. */
    BundleContents?: string[];
    /**
     * Unique identifier for the parent inventory item, as defined in the catalog, for object which were added from a bundle or
     * container.
     */
    BundleParent?: string;
    /** Catalog version for the inventory item, when this instance was created. */
    CatalogVersion?: string;
    /**
     * A set of custom key-value pairs on the instance of the inventory item, which is not to be confused with the catalog
     * item's custom data.
     */
    CustomData?: { [key: string]: string | null };
    /** CatalogItem.DisplayName at the time this item was purchased. */
    DisplayName?: string;
    /** Timestamp for when this instance will expire. */
    Expiration?: string;
    /** Class name for the inventory item, as defined in the catalog. */
    ItemClass?: string;
    /** Unique identifier for the inventory item, as defined in the catalog. */
    ItemId?: string;
    /** Unique item identifier for this specific instance of the item. */
    ItemInstanceId?: string;
    /** Timestamp for when this instance was purchased. */
    PurchaseDate?: string;
    /** Total number of remaining uses, if this is a consumable item. */
    RemainingUses?: number;
    /** Currency type for the cost of the catalog item. Not available when granting items. */
    UnitCurrency?: string;
    /** Cost of the catalog item in the given currency. Not available when granting items. */
    UnitPrice: number;
    /** The number of uses that were added or removed to this item in this call. */
    UsesIncrementedBy?: number;

}
// ...
export interface LoginResult extends PlayFabModule.IPlayFabResultCommon  {
    /**
     * If LoginTitlePlayerAccountEntity flag is set on the login request the title_player_account will also be logged in and
     * returned.
     */
    EntityToken?: EntityTokenResponse;
    /** Results for requested info. */
    InfoResultPayload?: GetPlayerCombinedInfoResultPayload;
    /** The time of this user's previous login. If there was no previous login, then it's DateTime.MinValue */
    LastLoginTime?: string;
    /** True if the account was newly created on this login. */
    NewlyCreated: boolean;
    /** Player's unique PlayFabId. */
    PlayFabId?: string;
    /** Unique token authorizing the user and game at the server level, for the current session. */
    SessionTicket?: string;
    /** Settings specific to this user. */
    SettingsForUser?: UserSettings;
    /** The experimentation treatments for this user at the time of login. */
    TreatmentAssignment?: TreatmentAssignment;

}
// ...
export interface PlayerProfileModel {
    /** List of advertising campaigns the player has been attributed to */
    AdCampaignAttributions?: AdCampaignAttributionModel[];
    /** URL of the player's avatar image */
    AvatarUrl?: string;
    /** If the player is currently banned, the UTC Date when the ban expires */
    BannedUntil?: string;
    /** List of all contact email info associated with the player account */
    ContactEmailAddresses?: ContactEmailInfoModel[];
    /** Player record created */
    Created?: string;
    /** Player display name */
    DisplayName?: string;
    /**
     * List of experiment variants for the player. Note that these variants are not guaranteed to be up-to-date when returned
     * during login because the player profile is updated only after login. Instead, use the LoginResult.TreatmentAssignment
     * property during login to get the correct variants and variables.
     */
    ExperimentVariants?: string[];
    /** UTC time when the player most recently logged in to the title */
    LastLogin?: string;
    /** List of all authentication systems linked to this player account */
    LinkedAccounts?: LinkedPlatformAccountModel[];
    /** List of geographic locations from which the player has logged in to the title */
    Locations?: LocationModel[];
    /** List of memberships for the player, along with whether are expired. */
    Memberships?: MembershipModel[];
    /** Player account origination */
    Origination?: string;
    /** PlayFab player account unique identifier */
    PlayerId?: string;
    /** Publisher this player belongs to */
    PublisherId?: string;
    /** List of configured end points registered for sending the player push notifications */
    PushNotificationRegistrations?: PushNotificationRegistrationModel[];
    /** List of leaderboard statistic values for the player */
    Statistics?: StatisticModel[];
    /** List of player's tags for segmentation */
    Tags?: TagModel[];
    /** Title ID this player profile applies to */
    TitleId?: string;
    /**
     * Sum of the player's purchases made with real-money currencies, converted to US dollars equivalent and represented as a
     * whole number of cents (1/100 USD). For example, 999 indicates nine dollars and ninety-nine cents.
     */
    TotalValueToDateInUSD?: number;
    /** List of the player's lifetime purchase totals, summed by real-money currency */
    ValuesToDate?: ValueToDateModel[];

}
// ...
export interface StatisticValue {
    /** unique name of the statistic */
    StatisticName?: string;
    /** statistic value for the player */
    Value: number;
    /** for updates to an existing statistic value for a player, the version of the statistic when it was loaded */
    Version: number;

}
// ...
export interface TreatmentAssignment {
    /** List of the experiment variables. */
    Variables?: Variable[];
    /** List of the experiment variants. */
    Variants?: string[];

}
// ...
export interface UserAccountInfo {
    /** User Android device information, if an Android device has been linked */
    AndroidDeviceInfo?: UserAndroidDeviceInfo;
    /** Sign in with Apple account information, if an Apple account has been linked */
    AppleAccountInfo?: UserAppleIdInfo;
    /** Timestamp indicating when the user account was created */
    Created: string;
    /** Custom ID information, if a custom ID has been assigned */
    CustomIdInfo?: UserCustomIdInfo;
    /** User Facebook information, if a Facebook account has been linked */
    FacebookInfo?: UserFacebookInfo;
    /** Facebook Instant Games account information, if a Facebook Instant Games account has been linked */
    FacebookInstantGamesIdInfo?: UserFacebookInstantGamesIdInfo;
    /** User Gamecenter information, if a Gamecenter account has been linked */
    GameCenterInfo?: UserGameCenterInfo;
    /** User Google account information, if a Google account has been linked */
    GoogleInfo?: UserGoogleInfo;
    /** User iOS device information, if an iOS device has been linked */
    IosDeviceInfo?: UserIosDeviceInfo;
    /** User Kongregate account information, if a Kongregate account has been linked */
    KongregateInfo?: UserKongregateInfo;
    /** Nintendo Switch account information, if a Nintendo Switch account has been linked */
    NintendoSwitchAccountInfo?: UserNintendoSwitchAccountIdInfo;
    /** Nintendo Switch device information, if a Nintendo Switch device has been linked */
    NintendoSwitchDeviceIdInfo?: UserNintendoSwitchDeviceIdInfo;
    /** OpenID Connect information, if any OpenID Connect accounts have been linked */
    OpenIdInfo?: UserOpenIdInfo[];
    /** Unique identifier for the user account */
    PlayFabId?: string;
    /** Personal information for the user which is considered more sensitive */
    PrivateInfo?: UserPrivateAccountInfo;
    /** User PSN account information, if a PSN account has been linked */
    PsnInfo?: UserPsnInfo;
    /** User Steam information, if a Steam account has been linked */
    SteamInfo?: UserSteamInfo;
    /** Title-specific information for the user account */
    TitleInfo?: UserTitleInfo;
    /** User Twitch account information, if a Twitch account has been linked */
    TwitchInfo?: UserTwitchInfo;
    /** User account name in the PlayFab service */
    Username?: string;
    /** User XBox account information, if a XBox account has been linked */
    XboxInfo?: UserXboxInfo;

}

export interface UserAndroidDeviceInfo {
    /** Android device ID */
    AndroidDeviceId?: string;

}

export interface UserAppleIdInfo {
    /** Apple subject ID */
    AppleSubjectId?: string;

}

export interface UserCustomIdInfo {
    /** Custom ID */
    CustomId?: string;

}

type UserDataPermission = "Private"
    | "Public";

export interface UserDataRecord {
    /** Timestamp for when this data was last updated. */
    LastUpdated: string;
    /**
     * Indicates whether this data can be read by all users (public) or only the user (private). This is used for GetUserData
     * requests being made by one player about another player.
     */
    Permission?: string;
    /** Data stored for the specified user data key. */
    Value?: string;

}
// ...
export interface UserSettings {
    /** Boolean for whether this player is eligible for gathering device info. */
    GatherDeviceInfo: boolean;
    /** Boolean for whether this player should report OnFocus play-time tracking. */
    GatherFocusInfo: boolean;
    /** Boolean for whether this player is eligible for ad tracking. */
    NeedsAttribution: boolean;

}
// ...
export interface Variable {
    /** Name of the variable. */
    Name: string;
    /** Value of the variable. */
    Value?: string;

}